[Front] selection of automatic entity types in Settings (remaining types errors) (#2417)
