+++ opencti-platform/opencti-front/src/utils/Localization.js
@@ -812,7 +812,9 @@ const i18n = {
       'Validate and send for import': 'Validar y enviar para importación',
       'Validate this workbench': 'Validar este espacio de trabajo',
       'File successfully removed': 'Fichero borrado con éxito',
+      'File and associated external reference successfully removed': 'Archivo y referencia externa asociada eliminados correctamente',
       'File successfully uploaded': 'Fichero subido con éxito',
+      'File successfully uploaded and associated external reference successfully created': 'Archivo cargado correctamente y referencia externa asociada creada correctamente',
       'Launch an import of this file': 'Lanzar importación de este fichero',
       Launch: 'Lanzar',
       'Rule rescan': 'Reescaneo de reglas',
@@ -2445,7 +2447,9 @@ const i18n = {
       'Validate and send for import': 'Valider et envoyer pour import',
       'Validate this workbench': 'Valider cet espace de travail',
       'File successfully removed': 'Fichier supprimé avec succès',
+      'File and associated external reference successfully removed': 'Fichier et référence externe associé supprimés avec succès',
       'File successfully uploaded': 'Fichier uploadé avec succès',
+      'File successfully uploaded and associated external reference successfully created': 'Fichier uploadé avec succès et référence externe associée créee avec succès',
       'Launch an import of this file': 'Lancer un import de ce fichier',
       Launch: 'Lancer',
       'Rule rescan': 'Rafraichissement des règles',
@@ -4034,7 +4038,9 @@ const i18n = {
       'Validate and send for import': 'インポート用に検証して送信する',
       'Validate this workbench': 'このワークスペースを検証する',
       'File successfully removed': 'ファイルの削除に成功',
+      'File and associated external reference successfully removed': 'ファイルおよび関連する外部参照が正常に削除されました',
       'File successfully uploaded': 'ファイルのアップロードに成功',
+      'File successfully uploaded and associated external reference successfully created': 'ファイルが正常にアップロードされ、関連付けられた外部参照が正常に作成されました',
       'Launch an import of this file': 'このファイルのインポートを開始する',
       Launch: '起動',
       'Rule rescan': 'ルールの再スキャン',
@@ -5553,7 +5559,9 @@ const i18n = {
       'Validate and send for import': '验证并发送以进行导入',
       'Validate this workbench': '验证此工作区',
       'File successfully removed': '已成功移除文件',
+      'File and associated external reference successfully removed': '已成功刪除檔和關聯的外部引用',
       'File successfully uploaded': '已成功上传文件',
+      'File successfully uploaded and associated external reference successfully created': '已成功上載檔並成功創建關聯的外部引用',
       'Launch an import of this file': '启动此文件的导入',
       Launch: '启动',
       'Rule rescan': '规则重新扫描',
