+++ opencti-platform/opencti-graphql/src/schema/internalObject.js
@@ -92,6 +92,7 @@ const internalObjectsAttributes = {
     'platform_login_message',
     'platform_enable_reference',
     'platform_hidden_types',
+    'platform_automatic_types',
     'created_at',
     'i_created_at_day',
     'i_created_at_month',
